requirements.txt
streamlit==1.34.0
fastapi==0.110.0
uvicorn[standard]==0.25.0
sqlmodel==0.0.19
pydantic==1.10.9
python-dotenv==1.0.1
openai==1.30.1
PyYAML==6.0.1
markdown==3.6 